Fix all unit tests build errors.

+++ src/resources.c
@@ -30,23 +30,29 @@
  * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
+/* stddef.h:
+ *   - size_t
+ */
+#include <stddef.h>
+
+
 #include "base.h"
 #include "resources.h"
 
 /* ---------------------------------------------------------------- */
 /* resource/tmp.txt.res */
 
-unsigned int get_binary_resource_tmp_txt_res_size(void)
+size_t get_binary_resource_tmp_txt_res_size(void)
 {
-  return (unsigned int)(((char *)(_binary_resource_tmp_txt_res_start)) - ((char *)(_binary_resource_tmp_txt_res_end)));
+  return (size_t)(((char *)(_binary_resource_tmp_txt_res_start)) - ((char *)(_binary_resource_tmp_txt_res_end)));
 }
 
 
-const char   *res_tmp_txt_start(void) { return _binary_resource_tmp_txt_res_start; }
-unsigned int  res_tmp_txt_size(void)  { return get_binary_resource_tmp_txt_res_size(); }
-const char   *res_tmp_txt_end(void)   { return _binary_resource_tmp_txt_res_end; }
+const char *res_tmp_txt_start(void) { return _binary_resource_tmp_txt_res_start; }
+size_t      res_tmp_txt_size(void)  { return get_binary_resource_tmp_txt_res_size(); }
+const char *res_tmp_txt_end(void)   { return _binary_resource_tmp_txt_res_end; }
 
-void res_tmp_txt(const char **start, unsigned int *size, const char **end)
+void res_tmp_txt(const char **start, size_t *size, const char **end)
 {
   if(start) *start = res_tmp_txt_start();
   if(size)  *size  = res_tmp_txt_size();
